Moved background Component; Updated Loader styles
