+++ src/Root/index.tsx
@@ -3,7 +3,7 @@ import App from "./App";
 import Landing from "./Landing";
 import { HashRouter } from "react-router-dom";
 import { loadTokenPrices } from "../helpers";
-import Loading from "../components/Loader";
+import Background from "../components/Background";
 
 function Root() {
     const isApp = (): boolean => {
@@ -22,6 +22,7 @@ function Root() {
     const app = () => (
         <HashRouter>
             <App />
+            <Background />
         </HashRouter>
     );
