+++ src/components/Background/background.scss
