+++ src/components/Loader/loader.scss
@@ -1,9 +1,29 @@
-.loader-wrap{
+.loader-wrap {
+    position: relative;
     width: 100%;
     height: 100vh;
     display: flex;
     justify-content: center;
     align-items: center;
-    background-color: #000000;
-    color: #FFFFFF;
+    color: #fff;
+    z-index: 20;
+}
+
+.loader-image {
+    width: 128px;
+    height: 128px;
+    overflow: hidden;
+    border-radius: 50%;
+    box-shadow: 0 0 60px 50px rgb(0 0 0 / 70%);
+    position: relative;
+
+    img {
+        width: 130px;
+        height: 130px;
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        transform: translate(-50%, -50%);
+        opacity: .8;
+    }
 }
