+++ src/components/ViewBase/index.tsx
@@ -62,15 +62,7 @@ function ViewBase({ children }: IViewBaseProps) {
                     <Drawer />
                 </Hidden>
             </div>
-     (
-                    <div className="App-Video">
-                     <video autoPlay muted loop className="App-Video-background">
-                        <source  src="background.mp4" type="video/mp4" />
-                </video>
-            </div>
-    ); 
             <div className={`${classes.content} ${isSmallerScreen && classes.contentShift}`}>
-            
                 {children}
             </div>
         </div>
