--- src/views/Landing/components/Background/index.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import "./background.scss";
-// import BlobsTop from "../../../../assets/icons/landing-blobs-top.png";
-
-function Background() {
-    return (
-        <div className="landing-background">
-            <div className="landing-background-blobs-top">
-                <video autoPlay muted loop id="myVideo">
-                    <source  src="background.mp4" type="video/mp4" />
-                </video>
-            </div>
-        </div>
-    );
-}
-
-export default Background;
