+++ src/views/Landing/components/Main/main.scss
@@ -3,7 +3,7 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    justify-content: end;
+    justify-content: flex-end;
     z-index: 2;
     position: relative;
