+++ src/views/Landing/index.tsx
@@ -3,7 +3,7 @@ import "./landing.scss";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import Main from "./components/Main";
-import Background from "./components/Background";
+import Background from "../../components/Background";
 
 function Landing() {
     return (
